Rename wormholeClient
