+++ libs/src/services/ethContracts/EthContracts.ts
@@ -8,7 +8,7 @@ import { StakingProxyClient } from './StakingProxyClient'
 import { DelegateManagerClient } from './DelegateManagerClient'
 import { ClaimsManagerClient } from './ClaimsManagerClient'
 import { ClaimDistributionClient } from './ClaimDistributionClient'
-import { WormholeClient } from './wormholeClient'
+import { WormholeClient } from './WormholeClient'
 import { EthRewardsManagerClient } from './EthRewardsManagerClient'
 import { TrustedNotifierManagerClient } from './TrustedNotifierManagerClient'
 import { Logger, Utils } from '../../utils'
